# backend/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "filex - files and agents API",
    "version": "0.50.0",
    "description": "The routes under `/api/files` (what the web explorer, the desktop app, embeds and `filex client` call) and `/api/ai` (the agent surface: REST and the MCP endpoint). Every operation here is a route the server registers: `web/tests/docs/openapiRoutes.test.ts` holds this file to `backend/internal/api/routes.go` in both directions - no operation the server does not serve, no `/api/files` or `/api/ai` route left out. The admin API (`/api/admin`, `/api/ai/admin`) is not described here; see docs/BACKEND.md.\n\nPaths of files are `<storage>://<path>`. A change sent by a browser with only the session cookie from another origin is refused (`403 cross_origin_refused`); a Bearer or `X-Filex-Token` header is not affected. An API token is held to its verbs (`x-filex-token-verb` on `/api/files`, `x-filex-token-scope` on `/api/ai`), and where an operation also asks a permission with a level, to that (`x-filex-token-permission`: `comments:read`, `comments:write` - the token's `comments:rw`; docs/RBAC.md).",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
...
  "servers": [
    {
      "url": "/",
      "description": "The filex server itself."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "tokenHeader": []
    },
    {
      "sessionCookie": []
    }
  ],
...
  "paths": {
...
    "/api/files/drafts/{key}/save": {
      "post": {
        "operationId": "saveDraft",
        "summary": "Put a draft in its folder",
        "tags": [
          "Drafts"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "as": {
                    "type": "string",
                    "description": "Save under this name instead (the `suggested` one of a taken answer)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "409": {
            "description": "`TARGET_TAKEN` with a `suggested` free name; nothing moved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "externalDocs": {
          "url": "https://github.com/BRF-Tech/filex/blob/main/docs/BACKEND.md#post-apifilesdraftskeysave"
        },
        "x-filex-token-verb": "write"
      }
    },
...
    "/api/files/plugins/ui/{plugin}/{view}/save": {
      "put": {
        "operationId": "appUiSave",
        "summary": "An app interface saves the file it opened",
        "tags": [
          "Apps"
        ],
        "parameters": [
          {
            "name": "plugin",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "view",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-filex-token-verb": "write"
      }
    },
...
    "/api/files/save-text": {
      "post": {
        "operationId": "saveText",
        "summary": "Save the built-in text editor's content",
        "tags": [
          "Files"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "path": {
                    "type": "string",
                    "description": "A storage path, `<storage>://<path>` (the storage root is `<storage>://`).",
                    "example": "docs://reports/q3.xlsx"
                  },
                  "content": {
                    "type": "string"
                  }
                },
                "required": [
                  "path",
                  "content"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "415": {
            "description": "Not a text file.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "externalDocs": {
          "url": "https://github.com/BRF-Tech/filex/blob/main/docs/BACKEND.md#post-apifilessave-text"
        },
        "x-filex-token-verb": "write"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A session token (`POST /api/auth/login`) or an API token (`fxt_...`)."
      },
      "tokenHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Filex-Token",
        "description": "An API token."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "filex_session",
        "description": "The web app's session."
      }
    },
    "responses": {
      "Error": {
        "description": "A refusal or a failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "The refusal's code: lower-case, stable across releases (`read_only`, `quota_exceeded`, `permission_denied`, ...). Some older routes still put an English sentence here; branch on the code, print `message` (docs/API-ERRORS.md)."
          },
          "code": {
            "type": "string",
            "description": "A second, upper-case code some routes have always sent beside `error` (`NAME_TAKEN`, `READ_ONLY`, `E2E_ENCRYPTED`, ...); kept for the clients that read it."
          },
          "message": {
            "type": "string",
            "description": "The sentence to show a person, written by the server in the reader's language (the account's, else Accept-Language). Print it as it is."
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "The values `message` was filled with (a name, a limit), for a program that wants them."
          },
          "hint": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
    }
  }
}